Add files via upload
Added a battery charging circuit underneath Analog Auto-Switching / Regulated Duty Cycles / Charge Batteries / Is this Sangulani Maxwell Chikumbutso's Method? Also, added a graphic to go with it invoked by the 'More Information' button in the upper right-hand corner of the simulator. And corrected its graphic. Some errors in my math!

To quote a YouTube video I recently posted on this latest circuit ...
It's far easier to rotate batteries out of one phase and into another so as to constantly have power for an electric car than it is to provide power directly to the motor while maintaining its batteries at a predefined voltage level. So, maybe, if one battery module is charged for just under 1 micro second, then allowed to rest for one percent of a micro second, then maybe it might be possible to discharge this battery into an electric motor for 28 micro seconds to complete its cycle of rotation? I don't know, but it sounds more doable than the former idea.

http://is.gd/sangulani
h