Add files via upload
Invented two phases of the life cycle of the same circuit:
Precharging it at the factory ...
http://is.gd/aerialpower

and taking it home in an EV or as an in-house power plant, etc ...
http://is.gd/freepower

This can deliver a constant level of amperage which is preset at the factory. So, this is a current source for an EV or building. It can accommodate any sized load of amperage.